Rename: app/api/types > lib/api/types 로 경로 변경
