+++ lib/api/types/error.ts
