+++ lib/api/types/room-service.dto.ts
