+++ lib/api/types/vote-service.dto.ts
